moved car logic
