+++ driver_service/src/main/java/com/internship/driver_service/controller/command/CommandProfileController.java
@@ -1,10 +1,8 @@
-package com.internship.driver_service.controller;
+package com.internship.driver_service.controller.command;
 
-import com.internship.driver_service.controller.doc.CommandDoc;
-import com.internship.driver_service.dto.CarDto;
+import com.internship.driver_service.controller.doc.ProfileCommandDoc;
 import com.internship.driver_service.dto.ProfileDto;
 import com.internship.driver_service.dto.RateDto;
-import com.internship.driver_service.service.command.CommandCarService;
 import com.internship.driver_service.service.command.CommandDriverProfileService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -19,10 +17,9 @@ import java.net.URI;
 @RestController
 @RequestMapping("/api/v1/drivers")
 @RequiredArgsConstructor
-public class CommandDriverProfileController implements CommandDoc {
+public class CommandProfileController implements ProfileCommandDoc {
 
     private final CommandDriverProfileService commandDriverProfileService;
-    private final CommandCarService commandCarService;
 
     @Override
     public ResponseEntity<ProfileDto> createProfile(@RequestBody ProfileDto profileDto) {
@@ -56,23 +53,6 @@ public class CommandDriverProfileController implements CommandDoc {
     @Override
     public ResponseEntity<Void> deleteRateFromDriver(@PathVariable Long id) {
         commandDriverProfileService.deleteRate(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @Override
-    public ResponseEntity<Void> setCurrentCar(@PathVariable Long id) {
-        commandCarService.setCurrentCar(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @Override
-    public CarDto addNewCar(@RequestBody CarDto carDto) {
-        return commandCarService.addNewCar(carDto);
-    }
-
-    @Override
-    public ResponseEntity<Void> deleteCar(@PathVariable Long id) {
-        commandCarService.deleteCar(id);
         return ResponseEntity.noContent().build();
     }
 }
