+++ driver_service/src/main/java/com/internship/driver_service/controller/doc/ProfileCommandDoc.java
@@ -1,6 +1,5 @@
 package com.internship.driver_service.controller.doc;
 
-import com.internship.driver_service.dto.CarDto;
 import com.internship.driver_service.dto.ProfileDto;
 import com.internship.driver_service.dto.RateDto;
 import com.internship.driver_service.utils.exceptions.transfer.BaseException;
@@ -18,7 +17,6 @@ import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.PositiveOrZero;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -29,8 +27,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
         name = "Operations for managing driver profiles",
         description = "Endpoints for creating, updating, and deleting driver profiles"
 )
-@RequestMapping("/api/v1/drivers")
-public interface CommandDoc {
+public interface ProfileCommandDoc {
 
     @Operation(
             summary = "Create a new driver profile",
@@ -157,82 +154,6 @@ public interface CommandDoc {
     @DeleteMapping("/rates/{id}")
     ResponseEntity<Void> deleteRateFromDriver(
             @Parameter(description = "Unique identifier of the rate", example = "1", required = true)
-            @PositiveOrZero(message = "id.positive")
-            @Max(value = ValidationConstants.MAX_ID_VALUE, message = "id.maxValue")
-            @PathVariable Long id);
-
-    @Operation(
-            summary = "Set a car as the current car for a driver",
-            description = "Marks a specific car as the current car for the driver"
-    )
-    @ApiResponses({
-            @ApiResponse(
-                    responseCode = "204",
-                    description = "Car set as current successfully"
-            ),
-            @ApiResponse(
-                    responseCode = "404",
-                    description = "Occurs when the car is not found",
-                    content = @Content(schema = @Schema(implementation = BaseException.class))
-            ),
-            @ApiResponse(
-                    responseCode = "400",
-                    description = "Invalid input parameters",
-                    content = @Content(schema = @Schema(implementation = BaseException.class))
-            )
-    })
-    @PatchMapping("/cars/{id}")
-    ResponseEntity<Void> setCurrentCar(
-            @Parameter(description = "Unique identifier of the car", example = "1", required = true)
-            @PositiveOrZero(message = "id.positive")
-            @Max(value = ValidationConstants.MAX_ID_VALUE, message = "id.maxValue")
-            @PathVariable Long id);
-
-    @Operation(
-            summary = "Add a new car for a driver",
-            description = "Adds a new car to the list of cars associated with the driver"
-    )
-    @ApiResponses({
-            @ApiResponse(
-                    responseCode = "201",
-                    description = "Car added successfully",
-                    content = @Content(schema = @Schema(implementation = CarDto.class))
-            ),
-            @ApiResponse(
-                    responseCode = "400",
-                    description = "Invalid input data",
-                    content = @Content(schema = @Schema(implementation = BaseValidationException.class))
-            )
-    })
-    @PostMapping("/cars")
-    CarDto addNewCar(
-            @Parameter(description = "Car data to add", required = true)
-            @Valid
-            @RequestBody CarDto carDto);
-
-    @Operation(
-            summary = "Delete a car from a driver",
-            description = "Deletes a specific car associated with the driver"
-    )
-    @ApiResponses({
-            @ApiResponse(
-                    responseCode = "204",
-                    description = "Car deleted successfully"
-            ),
-            @ApiResponse(
-                    responseCode = "404",
-                    description = "Occurs when the car is not found",
-                    content = @Content(schema = @Schema(implementation = BaseException.class))
-            ),
-            @ApiResponse(
-                    responseCode = "400",
-                    description = "Invalid input parameters",
-                    content = @Content(schema = @Schema(implementation = BaseException.class))
-            )
-    })
-    @DeleteMapping("/cars/{id}")
-    ResponseEntity<Void> deleteCar(
-            @Parameter(description = "Unique identifier of the car", example = "1", required = true)
             @PositiveOrZero(message = "id.positive")
             @Max(value = ValidationConstants.MAX_ID_VALUE, message = "id.maxValue")
             @PathVariable Long id);
