+++ driver_service/src/main/resources/application.yaml
@@ -11,5 +11,8 @@ spring:
     hibernate:
       ddl-auto: ${JPA_HIBERNATE_DDL_AUTO}
     show-sql: ${JPA_SHOW_SQL}
+  messages:
+    basename: messages
+    encoding: UTF-8
 server:
   port: ${SERVER_PORT}
